# Speed up `searchsortedfirst`.