This implementation:
- replaces midpoint computation with binary shift (division by 2)
- starts with a slightly narrower range `[lo, hi + 1]` instead of `[lo-1, hi+1]`
- narrows the range slightly faster by bumping `lo` to `m+1` instead of `m`
Overall this results in faster performance according to 
```julia
using Base.Order
using Base.Sort

function mysearchsortedfirst(v::AbstractVector, x, lo::T, hi::T, o::Ordering)::keytype(v) where T<:Integer
    hi = hi + T(1)
	len = hi - lo
    @inbounds while len != 0
		half_len = len >> 1
        m = lo + half_len
        if lt(o, v[m], x)
            lo = m + 1
			len -= half_len + 1
        else
            hi = m
			len = half_len
        end
    end
    return lo
end

xs = sort(randn(10000))
x = rand(xs)

using BenchmarkTools

@benchmark searchsortedfirst(xs, x, firstindex(xs), lastindex(xs), Forward)
```
BenchmarkTools.Trial: 10000 samples with 916 evaluations.
 Range (min … max):  118.512 ns …  2.973 μs  ┊ GC (min … max): 0.00% … 93.91%
 Time  (median):     121.469 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   125.719 ns ± 58.368 ns  ┊ GC (mean ± σ):  1.07% ±  2.26%

  ▄▇█▇▇▆▅▅▄▄▃▃▃▃▂▂▂▂▂▁▁                                        ▂
  ██████████████████████▇████▇▇▆█▇▇▇▇▇▇▇▆▇▆▆▆▇▅▆▇▅▅▅▆▅▅▅▅▆▅▅▅▅ █
  119 ns        Histogram: log(frequency) by time       162 ns <

 Memory estimate: 32 bytes, allocs estimate: 2.
```julia
@benchmark mysearchsortedfirst(xs, x, firstindex(xs), lastindex(xs), Forward)
```
BenchmarkTools.Trial: 10000 samples with 923 evaluations.
 Range (min … max):  113.570 ns …  2.627 μs  ┊ GC (min … max): 0.00% … 93.76%
 Time  (median):     116.356 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   120.122 ns ± 43.933 ns  ┊ GC (mean ± σ):  0.75% ±  2.04%

  ▁▆█▇▆▆▅▅▄▄▃▃▃▂▃▂▂▂▂▁▁▁                                       ▂
  ████████████████████████▇▇▇▇▆▇▇▇▇▆▅▇▆▇▇▆▆▆▆▆▆▆▇▄▆▅▅▅▅▅▅▄▄▃▅▄ █
  114 ns        Histogram: log(frequency) by time       154 ns <

I also ran some tests
```julia
using Test

for t in xs
	@test mysearchsortedfirst(xs, x, firstindex(xs), lastindex(xs), Forward) == searchsortedfirst(xs, x, firstindex(xs), lastindex(xs), Forward)
end

@test mysearchsortedfirst(xs, xs[1] - 1, firstindex(xs), lastindex(xs), Forward) == 1
@test mysearchsortedfirst(xs, xs[lastindex(xs)] + 1, firstindex(xs), lastindex(xs), Forward) == lastindex(xs) + 1
```
and all of them passed.

Apologies for the poor formatting. I was doing these changes in my Pluto notebook and they looked much better there.