@oscardssmith 2022-07-24
I believe this looks reasonable. One thing I'd be interested in is if it is faster to fall back to a linear search below a certain size (8 or so?)