@ttsugriy 2022-07-24
@oscardssmith using hybrid search/sort strategies oftentimes improve performance, especially in case of linear search that is both cache and branch predictor friendly. The trick though is coming up with good heuristics and they are likely to depend on multiple factors, including size width and whether the location of the searched element, if its present at all. I think it's a good area for research which unfortunately doesn't fit into the scope of this change.