@oscardssmith 2022-07-24
try replacing `while len != 0` with `while !(len == 0)`. According to @keno, the problem is `searchsortedfirst` gets used by the compiler before  `!=` exists.