@ttsugriy 2022-07-24
Thanks, @oscardssmith Unfortunately looks like changing to `!(len == 0)` didn't help :(