@N5N3 2022-07-24
Both `Core.Complier..!=` and `Core.Complier..==` has been defined before we include `sort.jl`, we just didn't import them into `Core.Complier.Sort`.
You can use `len > 0` instead or add `using .Base: !=, ==` at the top of `sort.jl`.
BTW, it would be good to extend this PR to `searchsortedlast`