@ttsugriy 2022-07-24
Good catch, @N5N3! Thank you! I'd be happy to do similar optimization for the `searchsortedlast` in a separate PR if this one gets merged.