@ttsugriy 2022-07-25
Looks like most of the builds/tests are now passing. The failures don't look related but I'd appreciate if someone could confirm.

Also, a better formatted version of the benchmark and testing is available as a Pluto notebook - https://ttsugriy.github.io/perf-bits/faster_lower_bound.jl.html