@ttsugriy 2022-07-29
> Indentation is a bit all over the place, [#46151 (comment)](https://github.com/JuliaLang/julia/pull/46151#discussion_r928199925) doesn't look addressed.

@KristofferC which part of that comment is not addressed? I have switched to `>>>` and the `midpoint` is not used intentionally. I was editing code in the github editor so I'll see if what may be off with indentation.