@Seelengrab 2022-08-06
Not 100% sure what Kristoffer was referring to, possibly just that there's no need to remove `midpoint` in the first place, since it does exactly the same thing as what you replaced it with.