@ttsugriy 2022-08-06
@Seelengrab `midpoint` does more than what I'm proposing in this change, which is why I'm proposing it after all :)

As I've explained in the original comment and my linked article, `midpoint` does more work since it's defined as `lo + ((hi - lo) >>> 0x01)` and proposed changes just do `len >>> 0x01` which is operationally cheaper.