@LilithHafner 2022-08-14
@ttsugriy, sorry to show up late, but I don't see the advantage of this PR.

> * replaces midpoint computation with binary shift (division by 2)

As [mentioned previously ](https://github.com/JuliaLang/julia/pull/46151#discussion_r928199925), we already use bit shift to compute midpoint. This PR replaced the subtraction, bit shift, and addition with
Bit shift
```julia
half_len = len >>> 0x01
```
addition
```julia
m = lo + half_len
```
assignment
```julia
len -= half_len + 1
# or
len = half_len
```
and conditional subtraction
```julia
len -= half_len + 1
```
I don't see much advantage.

> * starts with a slightly narrower range `[lo, hi + 1]` instead of `[lo-1, hi+1]`
> * narrows the range slightly faster by bumping `lo` to `m+1` instead of `m`

The previous version used a semantic where the returned index is in `(lo-1, hi+1]` and this version uses a semantic where the returned index is in [lo, hi+1]. The only performance difference between these two is that the previous version rounded down in midpoint calculations and this one rounds up. This results in some minor regressions (e.g. `serachsortedfirst([1,2], 0, 1, 2)`) and some minor improvements (e.g. `serachsortedfirst([1,2], 3, 1, 2)`). There are an equal number of regressions and improvements.

> * Overall this results in faster performance according to

119 ns to 114 ns is a 4% speedup. I would typically attribute that to benchmarking noise. I re-ran your benchmarks and some others on my machine and got results ranging from 1.5% improvement to 9.3% regression which makes me think it is unlikely this PR was an improvement, unlikely but plausible it was a regression, and likely it had no performance impact. To run my benchmarks, I created a file called `pr46151.jl` with the following content:

```julia
using Base.Sort, Base.Order
using Base.Sort: midpoint

function searchsortedfirst_A(v::AbstractVector, x, lo::T, hi::T, o::Ordering)::keytype(v) where T<:Integer
    u = T(1)
    lo = lo - u
    hi = hi + u
    @inbounds while lo < hi - u
        m = midpoint(lo, hi)
        if lt(o, v[m], x)
            lo = m
        else
            hi = m
        end
    end
    return hi
end

function searchsortedfirst_B(v::AbstractVector, x, lo::T, hi::T, o::Ordering)::keytype(v) where T<:Integer
    hi = hi + T(1)
    len = hi - lo
    @inbounds while len != 0
        half_len = len >>> 0x01
        m = lo + half_len
        if lt(o, v[m], x)
            lo = m + 1
            len -= half_len + 1
        else
            hi = m
            len = half_len
        end
    end
    return lo
end

xs = sort(randn(10000))
x = rand(xs)

using BenchmarkTools

@btime searchsortedfirst_A(xs, x, firstindex(xs), lastindex(xs), Forward)
@btime searchsortedfirst_B(xs, x, firstindex(xs), lastindex(xs), Forward)

@btime searchsortedfirst_A($xs, $x, firstindex($xs), lastindex($xs), Forward)
@btime searchsortedfirst_B($xs, $x, firstindex($xs), lastindex($xs), Forward)

@btime searchsortedfirst_A($xs, y, firstindex($xs), lastindex($xs), Forward) setup = (y=rand($xs))
@btime searchsortedfirst_B($xs, y, firstindex($xs), lastindex($xs), Forward) setup = (y=rand($xs))
```
And ran this at the repl:
```julia-repl
julia> include("pr46151.jl")
[ Info: Precompiling BenchmarkTools [6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf]
  88.439 ns (2 allocations: 32 bytes)
  87.260 ns (2 allocations: 32 bytes)
  29.959 ns (0 allocations: 0 bytes)
  29.483 ns (0 allocations: 0 bytes)
  28.706 ns (0 allocations: 0 bytes)
  30.540 ns (0 allocations: 0 bytes)
9291

julia> include("pr46151.jl")
  88.917 ns (2 allocations: 32 bytes)
  90.951 ns (2 allocations: 32 bytes)
  28.827 ns (0 allocations: 0 bytes)
  31.517 ns (0 allocations: 0 bytes)
  28.844 ns (0 allocations: 0 bytes)
  29.434 ns (0 allocations: 0 bytes)
7984

julia> include("pr46151.jl")
  92.137 ns (2 allocations: 32 bytes)
  94.066 ns (2 allocations: 32 bytes)
  30.138 ns (0 allocations: 0 bytes)
  32.980 ns (0 allocations: 0 bytes)
  28.166 ns (0 allocations: 0 bytes)
  30.013 ns (0 allocations: 0 bytes)
1630
```

A more robust analysis would consider llvm and or native code, but I have a hard time interpreting those results.